chore: Update rumble model, object target flow

+++ com.rotovr.android.bleplugin/unitybleplugin/src/main/java/com/rotovr/unitybleplugin/model/RumbleModel.java
@@ -3,6 +3,6 @@ package com.rotovr.unitybleplugin.model;
 import java.io.Serializable;
 
 public class RumbleModel implements Serializable {
-    public int Duration;
+    public float Duration;
     public int Power;
 }
