+++ com.rotovr.sdk.package/Assets/com.rotovr.sdk/Runtime/Model/RumbleModel.cs
@@ -5,13 +5,13 @@ namespace RotoVR.SDK.Model
     [Serializable]
     public class RumbleModel
     {
-        public RumbleModel(int duration, int power)
+        public RumbleModel(float duration, int power)
         {
             Duration = duration;
             Power = power;
         }
 
-        public int Duration { get; }
+        public float Duration { get; }
         public int Power { get; }
     }
 }
